[refactor] Core folder structure
