+++ Controllers/PlatformController.cs
@@ -2,9 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 using System.Collections.ObjectModel;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
-using GamesAPI.Core.Models;
 using GamesAPI.Dtos;
 using GamesAPI.Services;
+using GamesAPI.Models;
 
 namespace PlatformsAPI.Controllers;
