+++ Controllers/SoftwareHouseController.cs
@@ -2,10 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 using System.Collections.ObjectModel;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
-using GamesAPI.Core.Models;
+using GamesAPI.Models;
 using GamesAPI.Dtos;
 using GamesAPI.Services;
-using GamesAPI.Core.Services;
 
 namespace GamesAPI.Controllers;
