+++ Core/Mapper/Dtos/User/UserDetailDto.cs
@@ -1,4 +1,6 @@
-namespace GamesAPI.Dtos;
+using GamesAPI.Dtos;
+
+namespace GamesAPI.Core.Dtos;
 
 public class UserDetailDto: UserDto {
     public List<DeveloperDto>? DeveloperDtos { get; set; }
