+++ Core/Mapper/Dtos/User/UserDto.cs
@@ -1,4 +1,4 @@
-namespace GamesAPI.Dtos;
+namespace GamesAPI.Core.Dtos;
 
 public class UserDto {
     public int Id { get; set; }
