+++ Mapper/Dtos/Platform/PlatformDetailDto.cs
