+++ Mapper/Dtos/SoftwareHouse/SoftwareHouseDetailDto.cs
