+++ Mapper/Profiles/PlatformProfile.cs
@@ -1,8 +1,8 @@
 using AutoMapper;
-using GamesAPI.Core.Models;
+using GamesAPI.Models;
 using GamesAPI.Dtos;
 
-namespace GamesAPI.Utilities;
+namespace GamesAPI.Profiles;
 class PlatformProfile : Profile {
     public PlatformProfile() {
         CreateMap<CreatePlatformDto, Platform>();
