+++ Models/Category.cs
@@ -1,16 +1,14 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace GamesAPI.Core.Models;
-public class SoftwareHouse {
+namespace GamesAPI.Models;
+public class Category {
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Id { get; set; }
 
     [Required]
     public required string Name {get; set;}
-
-    public List<Developer>? Developers {get; set;}
 
     public List<Game>? Games {get; set;}
 }
