+++ Services/ReviewService.cs
@@ -1,6 +1,6 @@
 using AutoMapper;
 using GamesAPI.Core.DataContexts;
-using GamesAPI.Core.Models;
+using GamesAPI.Models;
 using GamesAPI.Core.Services;
 using GamesAPI.Dtos;
 using Microsoft.EntityFrameworkCore;
